requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic>=2
python-dotenv
pandas
requests